+++ fpdf/fonts.py
@@ -4,15 +4,22 @@ Includes the definition of the character widths of all PDF standard fonts.
 """
 import re
 
+from bisect import bisect_left
 from collections import defaultdict
 from dataclasses import dataclass, replace
-from typing import List, Optional, Union
+from typing import List, Optional, Tuple, Union
 
 from fontTools import ttLib
+
+try:
+    import uharfbuzz as hb
+except ImportError:
+    hb = None
 
 from .drawing import DeviceGray, DeviceRGB
 from .enums import FontDescriptorFlags, TextEmphasis
 from .syntax import Name, PDFObject
+from .util import escape_parens
 
 
 @dataclass
@@ -61,6 +68,14 @@ class CoreFont:
         self.fontkey = fontkey
         self.emphasis = TextEmphasis.coerce(style)
 
+    def get_text_width(self, text, font_size_pt, _):
+        return (len(text), sum(self.cw[c] for c in text) * font_size_pt * 0.001)
+
+    # Disabling this check - method kept as is to have same method/signature on CoreConf and TTFFont:
+    # pylint: disable=no-self-use
+    def encode_text(self, text):
+        return f"({escape_parens(text)}) Tj"
+
     def __repr__(self):
         return f"CoreFont(i={self.i}, fontkey={self.fontkey})"
 
@@ -71,14 +86,19 @@ class TTFFont:
         "type",
         "name",
         "desc",
+        "glyph_ids",
+        "hbfont",
         "up",
         "ut",
         "cw",
         "ttffile",
         "fontkey",
         "emphasis",
+        "scale",
         "subset",
         "cmap",
+        "ttfont",
+        "missing_glyphs",
     )
 
     def __init__(self, fpdf, font_file_path, fontkey, style):
@@ -87,67 +107,219 @@ class TTFFont:
         self.ttffile = font_file_path
         self.fontkey = fontkey
 
-        font = ttLib.TTFont(self.ttffile, fontNumber=0, lazy=True)
+        # recalcTimestamp=False means that it doesn't modify the "modified" timestamp in head table
+        # if we leave recalcTimestamp=True the tests will break every time
+        self.ttfont = ttLib.TTFont(
+            self.ttffile, recalcTimestamp=False, fontNumber=0, lazy=True
+        )
 
-        scale = 1000 / font["head"].unitsPerEm
-        default_width = round(scale * font["hmtx"].metrics[".notdef"][0])
+        self.scale = 1000 / self.ttfont["head"].unitsPerEm
+        default_width = round(self.scale * self.ttfont["hmtx"].metrics[".notdef"][0])
 
         try:
-            cap_height = font["OS/2"].sCapHeight
+            cap_height = self.ttfont["OS/2"].sCapHeight
         except AttributeError:
-            cap_height = font["hhea"].ascent
+            cap_height = self.ttfont["hhea"].ascent
 
         # entry for the PDF font descriptor specifying various characteristics of the font
         flags = FontDescriptorFlags.SYMBOLIC
-        if font["post"].isFixedPitch:
+        if self.ttfont["post"].isFixedPitch:
             flags |= FontDescriptorFlags.FIXED_PITCH
-        if font["post"].italicAngle != 0:
+        if self.ttfont["post"].italicAngle != 0:
             flags |= FontDescriptorFlags.ITALIC
-        if font["OS/2"].usWeightClass >= 600:
+        if self.ttfont["OS/2"].usWeightClass >= 600:
             flags |= FontDescriptorFlags.FORCE_BOLD
 
         self.desc = PDFFontDescriptor(
-            ascent=round(font["hhea"].ascent * scale),
-            descent=round(font["hhea"].descent * scale),
-            cap_height=round(cap_height * scale),
+            ascent=round(self.ttfont["hhea"].ascent * self.scale),
+            descent=round(self.ttfont["hhea"].descent * self.scale),
+            cap_height=round(cap_height * self.scale),
             flags=flags,
             font_b_box=(
-                f"[{font['head'].xMin * scale:.0f} {font['head'].yMin * scale:.0f}"
-                f" {font['head'].xMax * scale:.0f} {font['head'].yMax * scale:.0f}]"
+                f"[{self.ttfont['head'].xMin * self.scale:.0f} {self.ttfont['head'].yMin * self.scale:.0f}"
+                f" {self.ttfont['head'].xMax * self.scale:.0f} {self.ttfont['head'].yMax * self.scale:.0f}]"
             ),
-            italic_angle=int(font["post"].italicAngle),
-            stem_v=round(50 + int(pow((font["OS/2"].usWeightClass / 65), 2))),
+            italic_angle=int(self.ttfont["post"].italicAngle),
+            stem_v=round(50 + int(pow((self.ttfont["OS/2"].usWeightClass / 65), 2))),
             missing_width=default_width,
         )
 
         # a map unicode_char -> char_width
         self.cw = defaultdict(lambda: default_width)
-        self.cmap = tuple(font.getBestCmap().keys())
+
+        # fonttools cmap = unicode char to glyph name
+        # saving only the keys we have a tuple with
+        # the unicode characters available on the font
+        self.cmap = tuple(self.ttfont.getBestCmap().keys())
+
+        # saving a list of glyph ids to char to allow
+        # subset by unicode (regular) and by glyph
+        # (shaped with harfbuz)
+        self.glyph_ids = {}
+
         for char in self.cmap:
             # take glyph associated to char
-            glyph = font.getBestCmap()[char]
+            glyph = self.ttfont.getBestCmap()[char]
 
             # take width associated to glyph
-            w = font["hmtx"].metrics[glyph][0]
+            w = self.ttfont["hmtx"].metrics[glyph][0]
 
             # probably this check could be deleted
             if w == 65535:
                 w = 0
 
-            self.cw[char] = round(scale * w + 0.001)  # ROUND_HALF_UP
+            self.cw[char] = round(self.scale * w + 0.001)  # ROUND_HALF_UP
+
+            self.glyph_ids[char] = self.ttfont.getGlyphID(glyph)
+
+        self.missing_glyphs = []
 
         # include numbers in the subset! (if alias present)
         # ensure that alias is mapped 1-by-1 additionally (must be replaceable)
-        sbarr = "\x00 "
+        sbarr = "\x00 \r\n"
         if fpdf.str_alias_nb_pages:
             sbarr += "0123456789"
             sbarr += fpdf.str_alias_nb_pages
 
-        self.name = re.sub("[ ()]", "", font["name"].getBestFullName())
-        self.up = round(font["post"].underlinePosition * scale)
-        self.ut = round(font["post"].underlineThickness * scale)
+        self.name = re.sub("[ ()]", "", self.ttfont["name"].getBestFullName())
+        self.up = round(self.ttfont["post"].underlinePosition * self.scale)
+        self.ut = round(self.ttfont["post"].underlineThickness * self.scale)
         self.emphasis = TextEmphasis.coerce(style)
-        self.subset = SubsetMap([ord(char) for char in sbarr])
+        self.subset = SubsetMap(self, [ord(char) for char in sbarr])
+
+    def __repr__(self):
+        return f"TTFFont(i={self.i}, fontkey={self.fontkey})"
+
+    def close(self):
+        self.ttfont.close()
+        self.hbfont = None
+
+    def get_text_width(self, text, font_size_pt, text_shaping_parms):
+        if text_shaping_parms:
+            return self.shaped_text_width(text, font_size_pt, text_shaping_parms)
+        return (len(text), sum(self.cw[ord(c)] for c in text) * font_size_pt * 0.001)
+
+    def shaped_text_width(self, text, font_size_pt, text_shaping_parms):
+        """
+        When texts are shaped, the length of a string is not always the sum of all individual character widths
+        This method will invoke harfbuzz to perform the text shaping and return the sum of "x_advance"
+        and "x_offset" for each glyph. This method works for "left to right" or "right to left" texts.
+        """
+        _, glyph_positions = self.perform_harfbuzz_shaping(
+            text, font_size_pt, text_shaping_parms
+        )
+        text_width = 0
+        for pos in glyph_positions:
+            text_width += (
+                round(self.scale * (pos.x_advance + pos.x_offset) + 0.001)
+                * font_size_pt
+                * 0.001
+            )
+        return (len(glyph_positions), text_width)
+
+    # Disabling this check - looks like cython confuses pylint:
+    # pylint: disable=no-member
+    def perform_harfbuzz_shaping(self, text, font_size_pt, text_shaping_parms):
+        """
+        This method invokes Harfbuzz to perform text shaping of the input string
+        """
+        if not hasattr(self, "hbfont"):
+            self.hbfont = hb.Font(hb.Face(hb.Blob.from_file_path(self.ttffile)))
+        self.hbfont.ptem = font_size_pt
+        buf = hb.Buffer()
+        buf.cluster_level = 1
+        buf.add_str("".join(text))
+        features = text_shaping_parms["features"]
+        if (
+            text_shaping_parms["direction"]
+            or text_shaping_parms["script"]
+            or text_shaping_parms["language"]
+        ):
+            buf.direction = text_shaping_parms["direction"]
+            buf.script = text_shaping_parms["script"]
+            buf.language = text_shaping_parms["language"]
+        else:
+            buf.guess_segment_properties()
+        hb.shape(self.hbfont, buf, features)
+        return buf.glyph_infos, buf.glyph_positions
+
+    def encode_text(self, text):
+        txt_mapped = ""
+        for char in text:
+            uni = ord(char)
+            # Instead of adding the actual character to the stream its code is
+            # mapped to a position in the font's subset
+            txt_mapped += chr(self.subset.pick(uni))
+        return f'({escape_parens(txt_mapped.encode("utf-16-be").decode("latin-1"))}) Tj'
+
+    def shape_text(self, text, font_size_pt, text_shaping_parms):
+        """
+        This method will invoke harfbuzz for text shaping, include the mapping code
+        of the glyphs on the subset and map input characters to the cluster codes
+        """
+        if len(text) == 0:
+            return []
+        glyph_infos, glyph_positions = self.perform_harfbuzz_shaping(
+            text, font_size_pt, text_shaping_parms
+        )
+        text_info = []
+
+        # Find cluster gaps
+        # Ex: text = "ABCD"
+        # glyph infos has cluster: 0, 2, 3 - it means A and B are together on the first glyph
+        # (ligature or substitution) - the glyph should have both unicodes and it should be translated
+        # properly on the CID to GID mapping
+        #
+        def get_cluster_from_text_index(cluster_list, index):
+            pos = bisect_left(cluster_list, index)
+            if pos == 0:
+                return cluster_list[0]
+            if pos == len(cluster_list) or cluster_list[pos] != index:
+                return cluster_list[pos - 1]
+            return cluster_list[pos]
+
+        cluster_list = list(sorted(int(gi.cluster) for gi in glyph_infos))
+        cluster_mapping = {}
+        for i in range(len(text)):
+            cl = get_cluster_from_text_index(cluster_list, i)
+            if cl in cluster_mapping:
+                cluster_mapping[cl].append(i)
+            else:
+                cluster_mapping[cl] = [i]
+
+        for cluster_seq, gi in enumerate(glyph_infos):
+            unicode = []
+            if gi.cluster in cluster_mapping:
+                unicode = [ord(text[i]) for i in cluster_mapping[gi.cluster]]
+                cluster_mapping.pop(gi.cluster)
+
+            gname = self.ttfont.getGlyphName(gi.codepoint)
+            gwidth = round(self.scale * self.ttfont["hmtx"].metrics[gname][0])
+            glyph = self.subset.get_glyph(
+                glyph=gi.codepoint,
+                unicode=tuple(unicode),
+                glyph_name=gname,
+                glyph_width=gwidth,
+            )
+            force_positioning = False
+            if (
+                gwidth != glyph_positions[cluster_seq].x_advance
+                or glyph_positions[cluster_seq].x_offset != 0
+                or glyph_positions[cluster_seq].y_offset != 0
+                or glyph_positions[cluster_seq].y_advance != 0
+            ):
+                force_positioning = True
+            text_info.append(
+                {
+                    "mapped_char": self.subset.pick_glyph(glyph),
+                    "x_advance": glyph_positions[cluster_seq].x_advance,
+                    "y_advance": glyph_positions[cluster_seq].y_advance,
+                    "x_offset": glyph_positions[cluster_seq].x_offset,
+                    "y_offset": glyph_positions[cluster_seq].y_offset,
+                    "force_positioning": force_positioning,
+                }
+            )
+        return text_info
 
 
 class PDFFontDescriptor(PDFObject):
@@ -175,6 +347,21 @@ class PDFFontDescriptor(PDFObject):
         self.font_name = None
 
 
+@dataclass(order=True, frozen=True)
+class Glyph:
+    """
+    This represents one glyph on the font
+    Unicode is a tuple because ligatures or character substitution
+    can map a sequence of unicode characters to a single glyph
+    """
+
+    __slots__ = ["glyph_id", "unicode", "glyph_name", "glyph_width"]
+    glyph_id: int
+    unicode: Tuple
+    glyph_name: str
+    glyph_width: int
+
+
 class SubsetMap:
     """Holds a mapping of used characters and their position in the font's subset
 
@@ -186,34 +373,74 @@ class SubsetMap:
     the lowest possible representation.
     """
 
-    def __init__(self, identities: List[int]):
+    def __init__(self, font: TTFFont, identities: List[int]):
         super().__init__()
         self._next = 0
+        self.font = font
 
         # sort list to ease deletion once _next
         # becomes higher than first reservation
         self._reserved = sorted(identities)
 
         # int(x) to ensure values are integers
-        self._map = {x: int(x) for x in self._reserved}
+        self._map = {}
+        for x in self._reserved:
+            glyph = self.get_glyph(unicode=x)
+            if glyph:
+                self._map[glyph] = int(x)
 
     def __len__(self):
         return len(self._map)
 
     def pick(self, unicode: int):
-        if not unicode in self._map:
+        glyph = self.get_glyph(unicode=unicode)
+        if glyph is None and unicode not in self.font.missing_glyphs:
+            self.font.missing_glyphs.append(unicode)
+        return self.pick_glyph(glyph)
+
+    def pick_glyph(self, glyph):
+        if (glyph) and (glyph not in self._map):
             while self._next in self._reserved:
                 self._next += 1
                 if self._next > self._reserved[0]:
                     del self._reserved[0]
-
-            self._map[unicode] = self._next
+            self._map[glyph] = self._next
             self._next += 1
-
-        return self._map.get(unicode)
+        return self._map.get(glyph)
 
     def dict(self):
         return self._map.copy()
+
+    def get_glyph(
+        self, glyph=None, unicode=None, glyph_name=None, glyph_width=None
+    ) -> Glyph:
+        if glyph:
+            return Glyph(glyph, tuple(unicode), glyph_name, glyph_width)
+        if isinstance(unicode, int) and unicode in self.font.glyph_ids.keys():
+            return Glyph(
+                self.font.glyph_ids[unicode],
+                tuple([unicode]),
+                self.font.ttfont.getBestCmap()[unicode],
+                self.font.cw[unicode],
+            )
+        if unicode == 0x00:
+            return Glyph(self.font.cmap[0], tuple([0x00]), ".notdef", 0)
+        return None
+
+    def get_glyph_by_id(self, cid) -> Glyph:
+        for glyph in self._map:
+            if glyph.glyph_id == cid:
+                return glyph
+        return None
+
+    def get_glyph_by_unicode(self, cid) -> Glyph:
+        for glyph in self._map:
+            if glyph.unicode[0] == cid:
+                return glyph
+        return None
+
+    def get_all_glyph_names(self):
+        return [glyph.glyph_name for glyph in self._map]
 
 
 # Standard fonts